refactor: api subjects
